Rename main file to index.js.
